requirements.txt
PySide6==6.5.1  # GUI framework
python-docx>=0.8.11  # For handling DOCX files
pandas  # Optional: for managing tabular data
markdown==3.7
json
csv
pickle
